+++ xrpl/broadcast.js
@@ -1,17 +1,10 @@
 const { Command } = require('commander');
-const { mainProcessor, printInfo, prompt, decodeTxBlob } = require('./utils');
+const { mainProcessor } = require('./utils');
 const { addBaseOptions, addSkipPromptOption } = require('./cli-utils');
 
 async function broadcast(_config, _wallet, client, _chain, options, args) {
     const { txBlob } = args;
-    const tx = decodeTxBlob(txBlob);
-    printInfo('Preparing to broadcast transaction', tx);
-
-    if (prompt(`Submit ${tx.TransactionType} transaction?`, options.yes)) {
-        process.exit(0);
-    }
-
-    await client.submitTx(txBlob);
+    await broadcastTxBlob(client, txBlob, options);
 }
 
 if (require.main === module) {
